implemented actors getAll and count

+++ src/main/java/com/alvaroe/peliculas/http_response/Response.java
@@ -20,7 +20,7 @@ public class Response {
 
     private Object data;
 
-    private Integer totalRecords;
+    private long totalRecords;
 
     @JsonProperty("PaginationData")
     private Map<String, Object> pagination;
